% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaEnvGenerate.R
\name{betaEnv.generate}
\alias{betaEnv.generate}
\title{Generating a baseline landscape of vector-human and human-vector contact rates}
\usage{
betaEnv.generate(bmean, bcorrelation, stim)
}
\arguments{
\item{bmean}{the mean human <-> vector contact rate across the landscape}

\item{bcorrelation}{the degree of correlation between human susceptibility and human <-> vector contact rate, must be in interval -1 to 1}

\item{stim}{a vector of the starting immunity levels in each pixel}
}
\description{
Generates a continuous landscape of human <-> vector contact rates given parameters of mean and variance of contact rate and the degree of spatial correlation with susceptibility
}
\details{
Between the three supplied parameters this controls the intensity of transmission, its spatial variability and its spatial structure (in particularlly whether transmission intensity is most intense in areas that have experience recent transmission, or areas that have not experience recent transmission), i.e. does the outbreak happen in a new area or an existign area of high intensity transmission? The implementation is stochastic and an approximation.
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
sero <- c(0.577, 0.659, 0.814)
startweek <- 40
stim <- stim.generate(sgdat, sero, startweek, sgpop, unipix)
par(mfrow = c(2, 1))
plot.state(stim, sgpop, unipix)
betaEnv <- betaEnv.generate(bmean = 8, bcorrelation = -1, stim)
plot.state(betaEnv, sgpop, unipix)
}
\keyword{contact}
\keyword{human}
\keyword{mosquito}
\keyword{rate}
