% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseBacktrackTime.R
\name{case.backtrack.time}
\alias{case.backtrack.time}
\title{Establishing model initial conditions}
\usage{
case.backtrack.time(casedat, startweek, unipix, stim, betaEnv, mospdeath,
  hometime = 0.2)
}
\arguments{
\item{casedat}{Data frame of the locations, numbers and timings (in weeks) of recent cases, see ?sgdat}

\item{startweek}{Time (in weeks) at which the model simulation is to begin}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{stim}{Starting susceptibiltiy landscape, see ?stim.generate}

\item{betaEnv}{Starting human <-> mosquito contact rate landscape, see ?betaEnv.generate}

\item{mospdeath}{single value between 0 and 1, the daily probability of mosquito mortality}

\item{hometime}{Parameter that controls spatial concentration of movement, defaults to 0.2}
}
\description{
Generates the starting values for each of the model states given reported cases at time = 0
}
\details{
Uses supplied relationships to work out how many Infectious, exposed, etc are needed over time to generate the supplied reported detected human cases at t= 0.
These are then used for model burin.
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
sero <- c(0.577, 0.659, 0.814)
startweek <- 40
stim <- stim.generate(sgdat, sero, startweek, unipix)
betaEnv <- betaEnv.generate(bmean = 8, blogvariance = log(2), bcorrelation = -1, stim, unipix)
mospdeath <- 0.1
startTable <- case.backtrack.time(sgdat, startweek, unipix, stim, betaEnv, mospdeath, hometime = 0.5)
}
\keyword{burnin}
\keyword{generation}
\keyword{time}
