% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdeathMNFunc.R
\name{Cdeath.MN.func}
\alias{Cdeath.MN.func}
\title{Nested mosquito death sampler (due to vector control)}
\usage{
Cdeath.MN.func(mos_Es, probs)
}
\arguments{
\item{mos_Es}{Matrix or vector, matrix = exposed but not yet infectious mosquitoes that have not yet completed their EIP, vector = infectious mosquitoes}

\item{probs}{vector, patch-specific daily probability mosqutio dies due to reactive vector control}
}
\description{
Only for use within multinom()
}
\details{
Returns a matrix of dead mosquitoes
}
\examples{
mos_Es = matrix(round(runif(100, 0, 10), 0), nrow = 10, ncol = 10)
probs <- c(rep(0.5, 5), rep(0, 5))
Cdeath.MN.func(mos_Es, probs)

# infectious mosquitoes
mos_Is = round(runif(10, 0, 10), 0)
Cdeath.MN.func(mos_Is, probs)
}
