% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENSpatialEnsemble.R
\name{DEN.spatial.ensemble}
\alias{DEN.spatial.ensemble}
\title{Run a multiple realisations of the spatial dengue model and ensemble them}
\usage{
DEN.spatial.ensemble(weekdates, fitdat, pastdat, unipix, pixdistmat,
  steprun, nruns, qtiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ncores = 1, paramsList = NULL)
}
\arguments{
\item{weekdates}{Two element vector of the start and end weeks of the simulation over which the mdoel will be evaluated over}

\item{fitdat}{Data frame of the locations, numbers and timings (in weeks) of cases to fit the model to, see ?sgdat}

\item{pastdat}{Data frame of the locations, numbers and timings (in weeks) of all cases in the dataset (is used to generate the starting immunity profile), see ?sgdat}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{pixdistmat}{A patch distance matrix, see example}

\item{steprun}{integer, number of days for which the simulation should run, excluding burn in period}

\item{nruns}{Integer, number of stochastic model runs over which to average over}

\item{qtiles}{vector of quantiles over which results should be summarised (one or more values between 0 and 1)}

\item{ncores}{number of cpu cores to run simulations over (in parallel). Uses snowfall for parallelisation. Specify cores > 1 (default = 1) for parallel processing}

\item{paramsList}{Optional parameter list. If not supplied returns to defaults, see tutorial for full parameter list, see ?model.run for full list and explanation of parameters}
}
\description{
Runs multiple versions of DEN.spatial and calculates confidence intervals
}
\details{
Returns a list with each component summarising the results across multiple model runs at a given quantile, e.g. 0.95, 0.5, 0.05, etc.
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
pixdistmat <- distm(cbind(unipix$x, unipix$y))
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
weekdates <- c(40, 92)
# run model ensemble with default parameters
denmod_sim <- DEN.spatial.ensemble(weekdates, sgdat, sgdat, unipix, pixdistmat, 365, 10)
}
\keyword{ensemble}
\keyword{model}
\keyword{simulation}
