% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENSpatial.R
\name{DEN.spatial}
\alias{DEN.spatial}
\title{Run a single realisation of the spatial dengue model}
\usage{
DEN.spatial(weekdates, fitdat, pastdat, sgpop, unipix, pixdistmat, steprun,
  paramsList = NULL)
}
\arguments{
\item{weekdates}{Two element vector of the start and end weeks of the simulation over which the model will be evaluated over}

\item{fitdat}{Data frame of the locations, numbers and timings (in weeks) of cases to fit the model to, see ?sgdat}

\item{pastdat}{Data frame of the locations, numbers and timings (in weeks) of all cases in the dataset (is used to generate the starting immunity profile), see ?sgdat}

\item{sgpop}{Raster containing the number of people residing in each patch, see ?sgpop}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{pixdistmat}{A patch distance matrix, see example}

\item{steprun}{integer, number of days for which the simulation should run, excluding burn in period}

\item{paramsList}{Optional parameter list. If not supplied returns to defaults, see tutorial for full parameter list, see ?model.run for full list and explanation of parameters}
}
\description{
Main wrapper function for running a single realisation of the dengue spatial model, may take a while to run depending on specification
}
\details{
This function undertakes three main processes: i) fills in parameters with default options if not supplied, ii) generates a human movement matrix between patches,
ii) runs the model simulation.
within "paramsList()". Missing parameters will return to their default values. This function returns a outputs in a three element list giving:
i) daily counts of each model state (S, I, R, etc) sumed across the whole landscape, ii) daily counts of each model state for every patch,
iii) A treatment log (of length = number of steps) detailing which patches were treated each day
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
pixdistmat <- distm(cbind(unipix$x, unipix$y))
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
weekdates <- c(40, 92)
# model run with default parameters
denmod_sim <- DEN.spatial(weekdates, sgdat, sgdat, sgpop, unipix, pixdistmat, 365)
# model run with reactive drug deployment
drugtreat = data.frame(EffCoverage = 0.9, Duration = 30, Radius = 1000, Delay = 0)
denmod_sim_drugs <- DEN.spatial(weekdates, sgdat, unipix, pixdistmat, 365, paramsList = list(drugtreat = drugtreat))
}
\keyword{model}
\keyword{simulation}
