% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIPMNFunc.R
\name{EIP.MN.func}
\alias{EIP.MN.func}
\title{Nested EIP completion sampler}
\usage{
EIP.MN.func(mos_Es, probs)
}
\arguments{
\item{mos_Es}{Matrix, exposed but not yet infectious mosquitoes that have not yet died}

\item{probs}{Function, giving the probability of completing EIP on a given day}
}
\description{
Only for use within multinom()
}
\details{
Returns a matrix of newly infectious mosquitoes
}
\examples{
mos_Es = matrix(round(runif(100, 0, 10), 0), nrow = 10, ncol = 10)
probs <- function(times) plnorm(times, meanlog = log(7), sdlog = 0.21)
EIP.MN.func(mos_Es, probs)
}
