% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalWeights.R
\docType{data}
\name{finalWeights}
\alias{finalWeights}
\title{Posterior distribution of fitted model parameters}
\format{A list of length four (each parameter), within each contains a data.frame composed of 10000 rows and 2 columns
\describe{
  \item{names}{parameter each posterior is related to}
  \item{val}{value of the parameter at that sample}
  \item{prob}{Likelihood of the data given the model with that parameter value}
  }}
\source{
{contact author for more detail on fitting}
}
\usage{
finalWeights
}
\description{
A list of data.frames detailing the likelihood of each parameter value
'
}
\keyword{datasets}
