% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDspots.R
\name{find.Dspots}
\alias{find.Dspots}
\title{Find detected spots}
\usage{
find.Dspots(newsingD, pixdistmat, radius)
}
\arguments{
\item{newsingD}{Matrix of newly detected dengue infections, row = day of infection when they were detected, col = the patch in which they were detected}

\item{pixdistmat}{A distance matrix between patches}

\item{radius}{The radius in meters around the index case's residence that is to be searched}
}
\description{
Based on the location of detected (symptomatic) infections and a specified radius, identify which patches aroudn their residence will be subject to control activities
}
\details{
Returns a numeric vector with patch IDs that fall within the specified radius of an index case
}
\examples{
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
pixdistmat <- distm(cbind(unipix$x, unipix$y))
newsingD <- matrix(sample(c(0, 1), nrow(unipix) * 2, prob = c(0.99, 0.01), T), nrow = 2)
radius = 1000
find.Dspots(newsingD, pixdistmat, radius)
}
\keyword{index}
\keyword{radius}
