% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumDrugTreatMN.R
\name{Hum.Drug.Treat.MN}
\alias{Hum.Drug.Treat.MN}
\title{Nested Human drug treatment sampler}
\usage{
Hum.Drug.Treat.MN(sing_I, probs)
}
\arguments{
\item{sing_I}{Matrix, Infectious humans, each row is days since becoming infectious}

\item{probs}{Function, giving the probability of completing EIP on a given day}
}
\description{
Only for use within multinom()
}
\details{
Returns a matrix of newly prophylactic drug treated humans
}
\examples{
sing_I = matrix(round(runif(100, 0, 10), 0), nrow = 10, ncol = 10)
probs <- c(rep(0.5, 5), rep(0, 5))
Hum.Drug.Treat.MN(sing_I, probs)
}
