% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEvalSet.R
\name{make.eval.set}
\alias{make.eval.set}
\title{Creating an evaluation dataset for fitting}
\usage{
make.eval.set(fitdat, weekdates, unipix)
}
\arguments{
\item{fitdat}{Casedata, see ?sgdat for format}

\item{weekdates}{Two element vector of the start and end weeks of the simulation over which the mdoel will be evaluated over}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}
}
\description{
Processes the case data to generate a list of metrics required for different aspects of evaluating model fit
}
\details{
returns a list with three elements: i) weekly detialed number of cases per patch,
ii) Moran's weights for each pixel (for calculating Morans I)
iii) Weekly cumulative total cases
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
EvalSet <- make.eval.set(sgdat, weekdates = c(40, 92), unipix)
summary(EvalSet)
}
\keyword{evaluation}
