% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUnipix.R
\name{make.unipix}
\alias{make.unipix}
\title{Create a universal lookup table from a population raster}
\usage{
make.unipix(popras)
}
\arguments{
\item{popras}{A population raster, see ?sgpop}
}
\description{
This function takes takes a raster of population and creates a data.frame with key details including unique identifiers of the pixels and their characteristics including Longitude and Latitude and population.
This is used later on by the model for reference
}
\examples{
data(sgpop)
unipix <- make.unipix(sgpop)
}
\keyword{unipix}
