% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRun.R
\name{model.run}
\alias{model.run}
\title{Run a model simulation (workhorse function)}
\usage{
model.run(sing_S, sing_I, sing_R, sing_Rt, mos_S, mos_E, mos_I, betaEnv,
  mm_list2, mm_list_cum, pdetec, func_IIP, func_EIP, mospdeath,
  drugtreat_Eff, drugtreat_Dur, drugtreat_Rad, vectreat_Eff, vectreat_Dur,
  vectreat_Rad, steps = 365, unipix, pixdistmat, StopCriteria)
}
\arguments{
\item{sing_S}{vector, Susceptible human individuals in each pixel}

\item{sing_I}{matrix, Infectious human individuals in each pixel, rows = day of infection}

\item{sing_R}{vector, Recovered human individuals in each pixel}

\item{sing_Rt}{matrix, Recovered (temporary) human individuals in each pixel, rows = day of protection}

\item{mos_S}{vector, Susceptible mosquitoes in each pixel}

\item{mos_E}{matrix, Exposed but not yet infectious mosquitoes in each pixel, rows = day since infection}

\item{mos_I}{vector, Infectious mosquitoes in each pixel}

\item{betaEnv}{Starting human <-> mosquito contact rate landscape, see ?betaEnv.generate}

\item{mm_list2}{List of human movement matrices adjusted by the hometime parameter, see ?move.matrix.gene and ?DEN.spatial}

\item{mm_list_cum}{Matrix of patch effective population size (including both resident population and those who move there during the day)}

\item{pdetec}{Function for sampling probability of detecting a denge ifected individual (by passive surveillance), beta distributied probability}

\item{func_IIP}{Function for probability of completing human Intrinsic Incubation Period (time between infection and becoming symptomatic), a funciton of time since infection}

\item{func_EIP}{Function for probability of completing mosquito Extrinsic Incubation Period (time between mosquito becoming infected and becoming infectious), a function of time since infection}

\item{mospdeath}{Function for daily probability of death of a mosquito, beta distributed hazard, exponential distributed survival}

\item{drugtreat_Eff}{Effective coverage of prophylactic drug}

\item{drugtreat_Dur}{Duration of prophylactic drug effectiveness}

\item{drugtreat_Rad}{Radius of deployment around an index case of prophylactic drug}

\item{vectreat_Eff}{Effective coverage of vector control}

\item{vectreat_Dur}{Not used}

\item{vectreat_Rad}{Radius of deployment around an index case of vector control}

\item{steps}{Number of days for which the simulation shoudl run, defaults to 365}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{pixdistmat}{A patch distance matrix, see example in ?DENspatial}

\item{StopCriteria}{Single numeric value, if the model predicts reported cases exceed this value at any point during the simulation, simulation is terminated and "NA" is returned}

\item{mos_S}{vector, Susceptible mosquitoes in each pixel}

\item{drugtreat_Del}{Delay in days between detection of a dengue case and treatment of target population with drugs}

\item{vectreat_Del}{Delay in days between detection of a dengue case and treatment of target population with vector control}

\item{vectreat}{three element vector of details governing routine reactive vector control, see tutorial}
}
\description{
Directly implements the daily simulation of the dengue spatial model, usually used nested within a wrapper function, see ?DENSpatial
}
\details{
Implements a stochastic daily timestep model predictign the spatial spread of dengue across Singapore. As long as "StopCriteria" are not exceeded, will return a list of final values of each of the 
main stages of the model and three summaries: i) "totals" - population wide totals for each model state and ii) detailed_totals- pixel-specific totals for each model state, iii)
treatlog - log of length "steps" detailing which patches were treated with drugs or vector control on each day
}
\examples{
See tutorial
}
\keyword{model}
\keyword{simulation}
\keyword{workhorse}
