% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MosInfProb.R
\name{mos.inf.prob}
\alias{mos.inf.prob}
\title{Calculate probability of mosquito infection}
\usage{
mos.inf.prob(sing_I, betaEnv, unipix, mm_list2, mm_list_cum)
}
\arguments{
\item{sing_I}{matrix, Infectious human individuals in each pixel, rows = day of infection}

\item{betaEnv}{Starting human <-> mosquito contact rate landscape, see ?betaEnv.generate}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{mm_list2}{List of human movement matrices adjusted by the hometime parameter, see ?move.matrix.gene and ?DEN.spatial}

\item{mm_list_cum}{Matrix of patch effective population size (including both resident population and those who move there during the day)}
}
\description{
Calculates the pixel-specific probability of infection taking into account time allocation of infectious humans across the landscape
}
\details{
Calculates proabbility of infection of mosquitoes resident in each patch (mosquitoes don't move around) when infectious humans are moving around.
Infectiousness of humans is also dependent on their time since infection, peaking when symptomatic.
Returns a vector with length equal to the number of patches giving the proabbility of infection of each susceptible mosquito
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
sero <- c(0.577, 0.659, 0.814)
startweek <- 40
stim <- stim.generate(sgdat, sero, startweek, unipix)
betaEnv <- betaEnv.generate(bmean = 8, blogvariance = log(2), bcorrelation = -1, stim)
sing_I <- matrix(runif(15 * nrow(unipix), min = 0, max = 20), nrow = 15)
mlist <- move.matrix.gene(unipix, hometime = 0.2)
MIP <- mos.inf.prob(sing_I, betaEnv, unipix, mlist[[1]], mlist[[2]])
plot.state(MIP, sgpop, unipix)
}
\keyword{infection}
\keyword{mosquito}
\keyword{movement}
