% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosq.R
\docType{data}
\name{mosq}
\alias{mosq}
\title{Adult mosquito capture data for Singapore}
\format{A data frame with weekly values of GAI
\describe{
  \item{value}{The average number of adults per ovitrap across Singapore}
}}
\source{
\url{https://doi.org/10.1371/journal.pntd.0008528}
}
\usage{
mosq
}
\description{
A dataset containing Gravid Adult Index (GAI) for each week across Singapore
}
\keyword{datasets}
