% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoveMatrixGene.R
\name{move.matrix.gene}
\alias{move.matrix.gene}
\title{Creating a movement matrix}
\usage{
move.matrix.gene(unipix, hometime, type = "radiation")
}
\arguments{
\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{hometime}{Parameter that controls the spatial concentration of movement}

\item{type}{Human movement model type. Choose from  "radiation", "gravity" or "exponential"}
}
\description{
Applys a radiation model to a population surface to calcualte relative flux of human movement between patches
}
\details{
Returns a 2 element list: i) a list of length equal to the number of patches. Each item in this list is a movement matrix (in vector form) with fluxes between the home patch and all other patches, 
ii) a movement matrix (in vector form) that sums total time allocation in each pixel (from residents and visitors)
}
\examples{
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
mlist <- move.matrix.gene(unipix, hometime= 0.1, type = "radiation")
}
\keyword{movement}
