% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multinom.R
\name{multinom}
\alias{multinom}
\title{Multinomial state transition sampler}
\usage{
multinom(x, type)
}
\arguments{
\item{x}{a multi-dimentional list with level 1 giving the input and output processes and each element of level 1 containign a two element list with the starting state, e.g. sing_S and the probability of leaving that state}

\item{type}{a character indicating for which state ins and outs should be calculated for, e.g. "mos_E"}
}
\description{
Ensures arrivals to or departures from a model state have only one source / destination
}
\details{
Returns a list of new inputs and outputs from the specified state, maintaining the structure of each state.
}
\examples{
See tutorial
}
