% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatdeathMNFunc.R
\name{natdeath.MN.func}
\alias{natdeath.MN.func}
\title{Nested mosquito death sampler}
\usage{
natdeath.MN.func(mos_Es, probs)
}
\arguments{
\item{mos_Es}{Matrix, exposed but not yet infectious mosquitoes that have not yet completed their EIP}

\item{mospdeath}{single value between 0 and 1, the daily probability of mosquito mortality}
}
\description{
Only for use within multinom()
}
\details{
Returns a matrix of dead mosquitoes
}
\examples{
mos_Es = matrix(round(runif(100, 0, 10), 0), nrow = 10, ncol = 10)
probs <- function(x) rbeta(x, shape1 = 26.84987, shape2 = 107.3995)
natdeath.MN.func(mos_Es, probs)
}
