% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixIdFind.R
\name{pix.id.find}
\alias{pix.id.find}
\title{Match Longitude and Latitude of clusters to pixel IDs in grid}
\usage{
pix.id.find(longlat, unipix)
}
\arguments{
\item{longlat}{A two element vector of Longitude followed by Latitude}

\item{unipix}{A universal lookup table detailing all pixels within the grid. Must contain a pixel ID in the first column and the fields "x" (Longitude) and "y" (Latitude)}
}
\description{
This function takes latitude and longitude of identified cases and matched them to a specific identifier from the unipix dataframe
}
\details{
uses pixel stepwise distance (Longitude + Latitude) to identify nearest pixel to the given LatLong.
}
\examples{
data(sgdat)
data(sgpop)
unipix <- make.unipix(sgpop)
pix.id.find(c(sgdat$Longitude[1], sgdat$Latitude[1]), unipix)
}
\keyword{pixel}
