% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotState.R
\name{plot.state}
\alias{plot.state}
\title{Plots a map of a given model compartment state}
\usage{
\method{plot}{state}(state, examras, unipix, main = NULL)
}
\arguments{
\item{state}{Matrix or vector of a model state, e.g. S, I or R}

\item{examras}{Population raster used to determin the landscape, e.g. sgpop}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{main}{Optional character title}
}
\description{
Takes a vector or matrix of a model state, e.g. S, I or R, and plots it back on the original supplied population map
}
\examples{
data(sgdat)
seromedian <- 0.659
serolow <- 0.577
serohigh <- 0.814
startweek <- 40
stim <- Stim.generate(sgdat, seromedian, serolow, serohigh, startweek)
plot.state(stim)
}
\keyword{map}
\keyword{plot}
\keyword{state}
