% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopProcess.R
\name{pop.process}
\alias{pop.process}
\title{Population processing}
\usage{
pop.process(sgpop, agg = FALSE)
}
\arguments{
\item{sgpop}{A raster of human population counts}

\item{agg}{Multiplier for level of aggregation of the raster, defaults to FALSE (0)}
}
\description{
This function processes a raster of human population (counts) to be compatible with the dengue model
}
\details{
Processes performed involve 
i) making all pixels integer value (so each pixel contains a discrete number of people)
ii) if supplied "agg" determines the scale of pixel aggregation, e.g. agg = 10 will convert a raster with pixels sized 100m x 100m to 1000m x 1000m
}
\examples{
data(sgpop)
sgpop10 <- pop.process(sgpop, agg = 10)
par(mfrow = c(2, 1))
plot(sgpop)
plot(sgpop10)
}
\keyword{population}
