% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadModel.R
\name{rad.model}
\alias{rad.model}
\title{Radiation model}
\usage{
rad.model(m, n, s)
}
\arguments{
\item{m}{human population size in patch m (home patch)}

\item{n}{human population size in patch n}

\item{s}{human population size in all pixels within the radius m -> n, excluding m and n}
}
\description{
Human movement model based on radiation, for details see Simini et al. (2012) Nature. 484, 96-100
}
\details{
returns the predicted flux between pixels m and n
}
\examples{
popmat <- matrix(round(runif(16, 5, 20), 0), nrow = 4, ncol = 4)
m <- popmat[1, 1]
n <- popmat[3, 2]
distm <- as.matrix(dist(popmat)) # creates a distance matrix
r <- distm[3, 2]
s <- sum(popmat[distm <= r]) - m - n
rad.model(m, n, s)
}
\keyword{movement}
