% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_vector.R
\docType{data}
\name{seasonal_vector}
\alias{seasonal_vector}
\title{Daily incidence rate ratios}
\format{A vector composed of 365 elements
\describe{
  \item{V1}{Daily incidence rate ratio}
}}
\source{
\url{DOI: 10.1186/s12879-017-2326-8}
}
\usage{
seasonal_vector
}
\description{
Monthly incidence rate ratios of reported dengue cases in Khanh Hoa, Vietnam, between 1994 - 2003, were converetd to daily data for every day of the year.
}
\keyword{datasets}
