% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero.R
\docType{data}
\name{sero}
\alias{sero}
\title{Prevalence of dengue immunity in the singapore population}
\format{A georeferenced raster file with pixels aligned in a 375 row, 579 column grid:
\describe{
  \item{value}{prevalence of dengue immunity}
}}
\source{
\url{https://academic.oup.com/aje/article/188/8/1529/5486392}
}
\usage{
sero
}
\description{
An estimate of the prevalence of dengue immunity in Singapore based on the local age distribution of the population
}
\keyword{datasets}
