% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpop.R
\docType{data}
\name{sgpop}
\alias{sgpop}
\title{Population of Singapore in raster format}
\format{A georeferenced raster file with pixels aligned in a 184 row, 271 column grid:
\describe{
  \item{value}{number of people residing in pixel}
}}
\source{
\url{http://www.worldpop.org.uk}
}
\usage{
sgpop
}
\description{
A dataset containing the number of people residing in each 100m x 100m pixel in the nation of Singapore, adjusted to match the UN 2015 national population totals.
}
\keyword{datasets}
