% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StimGenerate.R
\name{stim.generate}
\alias{stim.generate}
\title{Generating a baseline landscape of susceptible individuals}
\usage{
stim.generate(pastdat, sero, startweek, unipix)
}
\arguments{
\item{pastdat}{Data frame of the locations, numbers and timings (in weeks) of recent cases, see ?sgdat}

\item{sero}{a three element vector giving the minimum, median and maximum observed seroprevalence in the landscape}

\item{startweek}{the week from which the model simulation will begin}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}
}
\description{
Generates a continuous landscape of susceptible individuals given key parameters on the median and range of dengue seroprevalence and the past distribution of recent cases
}
\details{
This approach assumes that true seroprevalence will be higher in areas with recent dengue infection. It therefore, takes past data from before the model simulation, uses krigeing (see ?autoKrig) to generate a continuosu surface of predicted past cases, then scales these to match the range and median of the seroprevalence values provided.
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
sero <- c(0.577, 0.659, 0.814)
startweek <- 40
stim <- stim.generate(sgdat, sero, startweek, unipix)
plot.state(stim, sgpop, unipix)
}
\keyword{krige}
\keyword{landscape}
\keyword{susceptible}
