% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{transition}
\alias{transition}
\title{Matrix time transition function}
\usage{
transition(base, plus, minus)
}
\arguments{
\item{base}{matrix, the model state, e.g. mos_E}

\item{plus}{vector, new additions to the state, e.g. newmosE}

\item{minus}{matrix, new subtractions from the state, e.g. newmosD}
}
\description{
As rows indicate time in each state, at the end of the timestep, each of the matrix-based states need to be moved down a row
}
\details{
Removes those who graduate from the state, shuffles down the rows by one then adds new additions to the first row of the matrix
}
\examples{
mos_Es = matrix(round(runif(100, 5, 10), 0), nrow = 10, ncol = 10)
newmosE = round(runif(10, 0, 2), 0)
newmosD = matrix(round(runif(100, 0, 1), 0), nrow = 10, ncol = 10)
mos_EsT2 <- transition(mos_Es, newmosE, newmosD)
}
