% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCTEffCalc.R
\name{UCT.Eff.calc}
\alias{UCT.Eff.calc}
\title{Summarises the results from multiple simulations of a UCT}
\usage{
UCT.Eff.calc(denmod_UCT, unipix)
}
\arguments{
\item{denmod_UCT}{A list of multiple SpatialDengue model objects}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}
}
\description{
Calculates Effectiveness and ttoal people treated from the outputs of multiple simulated trials
}
\details{
plots a boxplot of Effectiveness against dengue infection and dengue cases and returns a data frame of the Effectiveness and 
total person-days of observation in the trial and total people treated with drugs for each realisation of the trial (model run)
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
pixdistmat <- distm(cbind(unipix$x, unipix$y))
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
weekdates <- c(40, 92)

UCTinfo <- list(Tstart = 50,
               Tfollowup = 28,
               ntreat = 30,
               ncontrol = 30,
               DrugEfficacy = 0.9)

denmod_UCT = DEN.spatial.UCT(weekdates, sgdat, unipix, pixdistmat, UCTinfo, nruns = 10)
effectiveness <- UCT.Eff.calc(denmod_UCT, unipix)
}
\keyword{summary}
