% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotState.R
\name{plot.state}
\alias{plot.state}
\title{Plots a map of a given model compartment state}
\usage{
plot.state(state, examras, unipix, main = NULL)
}
\arguments{
\item{state}{Matrix or vector of a model state, e.g. S, I or R}

\item{examras}{Population raster used to determin the landscape, e.g. sgpop}

\item{unipix}{Universal pixel lookup table, see ?make.unipix}

\item{main}{Optional character title}
}
\description{
Takes a vector or matrix of a model state, e.g. S, I or R, and plots it back on the original supplied population map
}
\examples{
data(sgdat)
data(sgpop)
sgpop <- pop.process(sgpop, agg = 10)
unipix <- make.unipix(sgpop)
sgdat <- data.frame(sgdat, patchID = apply(cbind(sgdat[, 3:2]), 1, pix.id.find, unipix))
sero <- c(0.577, 0.659, 0.814)
startweek <- 40
stim <- stim.generate(sgdat, sero, startweek, sgpop, unipix)
plot.state(stim)
}
\keyword{map}
\keyword{plot}
\keyword{state}
