% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpop.R
\docType{data}
\name{sgpop}
\alias{sgpop}
\title{Population of Singapore in raster format}
\format{A georeferenced raster file with pixels aligned in a 375 row, 579 column grid:
\describe{
  \item{value}{number of people residing in pixel}
}}
\source{
\url{www.openstreetmap.org}
}
\usage{
sgpop
}
\description{
A dataset containing the number of people residing in each 100m x 100m pixel in the nation of Singapore created using Open Street Map data, adjusted to match 2013 census totals.
}
\keyword{datasets}
