% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpop.R
\docType{data}
\name{sgpop}
\alias{sgpop}
\title{Population of Singapore in raster format}
\format{A georeferenced raster file with pixels aligned in a 184 row, 271 column grid:
\describe{
  \item{value}{number of people residing in pixel}
}}
\source{
\url{https://www.openstreetmap.org/#map=6/54.910/-3.432}
}
\usage{
sgpop
}
\description{
A dataset containing the number of people residing in each 100m x 100m pixel in the nation of Singapore using openstreetmap building data.
}
\keyword{datasets}
